Improve testing of different Ruby, Rails versions
* Remove `Gemfile.lock` file from version control.
* Remove `gemfiles` from version control. Let Appraisal generate them.
* Use Semantic Versioning (http://semver.org/) in `Appraisal`.

This intends to solve the following problems:

* The `Gemfile.lock` and `gemfiles/*.gemfile.lock` files were being
  updated regularly during normally development work, polluting commits
  and code reviews with irrelevant changes like `tzinfo` changing from
  0.3.35 to 0.3.37.
* The pathnames in the appraised `.lock` files change every time a new
  developer runs the suite, polluting commits and code reviews.
* We needed to manually update patch versions in `Appraisal` to test
  against the latest versions, polluting the git history, instead of
  just triggering a new Travis build. This is the most important
  combination of dependencies and therefore should be quick and easy to
  test.

http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/

Also:

* Clarify versions of Ruby/Rails supported in README
* Loosen development dependency versions to 2-digit SemVer.
* Fix config.paths for Rails 4.
* Fix controller specs on Rails 4.
* Don't fork the Cucumber suite (leftover from Guard/Spork days?).
* Fix `Cucumber::Rails::Database` error:

    uninitialized constant Cucumber::Rails::Database (NameError)

* Use Capybara 2.0 spec/features convention.
* Add RSpec as an explicit dependency to avoid this errro:

    require': cannot load such file -- rspec (LoadError)

https://github.com/thoughtbot/shoulda-matchers/pull/262
