+++ .gitignore
@@ -1,11 +1,13 @@
-*.DS_Store
-*.swp
-*.swo
-*~
 !.keep
+*.DS_Store
+*.swo
+*.swp
+*~
 .bundle
+Gemfile.lock
+bin/
 db/*.sqlite3
+gemfiles/
 log/*.log
 pkg
 tmp/
-bin/
