+++ Appraisals
@@ -1,11 +1,9 @@
 if RUBY_VERSION >= '2.0'
-  rails_versions = ['3.2.13.rc2']
+  name, version = 'ruby.2.support', '>= 3.2.13'
 else
-  rails_versions = ['3.0.20', '3.1.11', '3.2.12']
+  name, version = 'rails.3.support', '~> 3.0'
 end
 
-rails_versions.each do |rails_version|
-  appraise "#{rails_version}" do
-    gem 'rails', rails_version
-  end
+appraise name do
+  gem 'rails', version
 end
