+++ clearance.gemspec
@@ -25,17 +25,17 @@ Gem::Specification.new do |s|
   s.add_dependency 'bcrypt-ruby'
   s.add_dependency 'email_validator'
   s.add_dependency 'rails', '>= 3.0'
-  s.add_development_dependency 'appraisal', '0.4.1'
-  s.add_development_dependency 'aruba', '0.4.11'
-  s.add_development_dependency 'bourne', '1.3.0'
-  s.add_development_dependency 'bundler', '~> 1.1'
-  s.add_development_dependency 'capybara', '1.1.2'
-  s.add_development_dependency 'cucumber-rails', '1.1.1'
-  s.add_development_dependency 'database_cleaner', '0.8.0'
-  s.add_development_dependency 'factory_girl_rails', '3.5.0'
-  s.add_development_dependency 'psych', '~> 1.3.4'
-  s.add_development_dependency 'rspec-rails', '2.12.2'
-  s.add_development_dependency 'shoulda-matchers', '1.2.0'
-  s.add_development_dependency 'sqlite3', '1.3.6'
-  s.add_development_dependency 'timecop', '0.3.5'
+  s.add_development_dependency 'appraisal', '~> 0.5'
+  s.add_development_dependency 'aruba', '~> 0.5'
+  s.add_development_dependency 'bundler', '~> 1.3'
+  s.add_development_dependency 'bourne', '~> 1.4'
+  s.add_development_dependency 'capybara', '~> 2.0'
+  s.add_development_dependency 'cucumber-rails', '~> 1.3'
+  s.add_development_dependency 'database_cleaner', '~> 0.9'
+  s.add_development_dependency 'factory_girl_rails', '~> 4.2'
+  s.add_development_dependency 'rspec', '~> 2.13'
+  s.add_development_dependency 'rspec-rails', '~> 2.13'
+  s.add_development_dependency 'shoulda-matchers', '~> 1.5'
+  s.add_development_dependency 'sqlite3', '~> 1.3'
+  s.add_development_dependency 'timecop', '~> 0.6'
 end
