--- gemfiles/3.0.20.gemfile
@@ -1,7 +1,0 @@
-# This file was generated by Appraisal
-
-source "http://rubygems.org"
-
-gem "rails", "3.0.20"
-
-gemspec :path=>"../"
