--- gemfiles/3.0.20.gemfile.lock
@@ -1,173 +1,0 @@
-PATH
-  remote: ../
-  specs:
-    clearance (1.0.0.rc5)
-      bcrypt-ruby
-      email_validator
-      rails (>= 3.0)
-
-GEM
-  remote: http://rubygems.org/
-  specs:
-    abstract (1.0.0)
-    actionmailer (3.0.20)
-      actionpack (= 3.0.20)
-      mail (~> 2.2.19)
-    actionpack (3.0.20)
-      activemodel (= 3.0.20)
-      activesupport (= 3.0.20)
-      builder (~> 2.1.2)
-      erubis (~> 2.6.6)
-      i18n (~> 0.5.0)
-      rack (~> 1.2.5)
-      rack-mount (~> 0.6.14)
-      rack-test (~> 0.5.7)
-      tzinfo (~> 0.3.23)
-    activemodel (3.0.20)
-      activesupport (= 3.0.20)
-      builder (~> 2.1.2)
-      i18n (~> 0.5.0)
-    activerecord (3.0.20)
-      activemodel (= 3.0.20)
-      activesupport (= 3.0.20)
-      arel (~> 2.0.10)
-      tzinfo (~> 0.3.23)
-    activeresource (3.0.20)
-      activemodel (= 3.0.20)
-      activesupport (= 3.0.20)
-    activesupport (3.0.20)
-    appraisal (0.4.1)
-      bundler
-      rake
-    arel (2.0.10)
-    aruba (0.4.11)
-      childprocess (>= 0.2.3)
-      cucumber (>= 1.1.1)
-      ffi (>= 1.0.11)
-      rspec (>= 2.7.0)
-    bcrypt-ruby (3.0.1)
-    bourne (1.3.0)
-      mocha (= 0.13.0)
-    builder (2.1.2)
-    capybara (1.1.2)
-      mime-types (>= 1.16)
-      nokogiri (>= 1.3.3)
-      rack (>= 1.0.0)
-      rack-test (>= 0.5.4)
-      selenium-webdriver (~> 2.0)
-      xpath (~> 0.1.4)
-    childprocess (0.3.7)
-      ffi (~> 1.0, >= 1.0.6)
-    cucumber (1.2.1)
-      builder (>= 2.1.2)
-      diff-lcs (>= 1.1.3)
-      gherkin (~> 2.11.0)
-      json (>= 1.4.6)
-    cucumber-rails (1.1.1)
-      capybara (>= 1.1.1)
-      cucumber (>= 1.1.0)
-      nokogiri (>= 1.5.0)
-    database_cleaner (0.8.0)
-    diff-lcs (1.1.3)
-    email_validator (1.3.0)
-      activemodel
-    erubis (2.6.6)
-      abstract (>= 1.0.0)
-    factory_girl (3.5.0)
-      activesupport (>= 3.0.0)
-    factory_girl_rails (3.5.0)
-      factory_girl (~> 3.5.0)
-      railties (>= 3.0.0)
-    ffi (1.3.1)
-    gherkin (2.11.5)
-      json (>= 1.4.6)
-    i18n (0.5.0)
-    json (1.7.6)
-    mail (2.2.19)
-      activesupport (>= 2.3.6)
-      i18n (>= 0.4.0)
-      mime-types (~> 1.16)
-      treetop (~> 1.4.8)
-    metaclass (0.0.1)
-    mime-types (1.19)
-    mocha (0.13.0)
-      metaclass (~> 0.0.1)
-    multi_json (1.5.0)
-    nokogiri (1.5.6)
-    polyglot (0.3.3)
-    psych (1.3.4)
-    rack (1.2.7)
-    rack-mount (0.6.14)
-      rack (>= 1.0.0)
-    rack-test (0.5.7)
-      rack (>= 1.0)
-    rails (3.0.20)
-      actionmailer (= 3.0.20)
-      actionpack (= 3.0.20)
-      activerecord (= 3.0.20)
-      activeresource (= 3.0.20)
-      activesupport (= 3.0.20)
-      bundler (~> 1.0)
-      railties (= 3.0.20)
-    railties (3.0.20)
-      actionpack (= 3.0.20)
-      activesupport (= 3.0.20)
-      rake (>= 0.8.7)
-      rdoc (~> 3.4)
-      thor (~> 0.14.4)
-    rake (10.0.3)
-    rdoc (3.12)
-      json (~> 1.4)
-    rspec (2.12.0)
-      rspec-core (~> 2.12.0)
-      rspec-expectations (~> 2.12.0)
-      rspec-mocks (~> 2.12.0)
-    rspec-core (2.12.2)
-    rspec-expectations (2.12.1)
-      diff-lcs (~> 1.1.3)
-    rspec-mocks (2.12.2)
-    rspec-rails (2.12.2)
-      actionpack (>= 3.0)
-      activesupport (>= 3.0)
-      railties (>= 3.0)
-      rspec-core (~> 2.12.0)
-      rspec-expectations (~> 2.12.0)
-      rspec-mocks (~> 2.12.0)
-    rubyzip (0.9.9)
-    selenium-webdriver (2.29.0)
-      childprocess (>= 0.2.5)
-      multi_json (~> 1.0)
-      rubyzip
-      websocket (~> 1.0.4)
-    shoulda-matchers (1.2.0)
-      activesupport (>= 3.0.0)
-    sqlite3 (1.3.6)
-    thor (0.14.6)
-    timecop (0.3.5)
-    treetop (1.4.12)
-      polyglot
-      polyglot (>= 0.3.1)
-    tzinfo (0.3.35)
-    websocket (1.0.7)
-    xpath (0.1.4)
-      nokogiri (~> 1.3)
-
-PLATFORMS
-  ruby
-
-DEPENDENCIES
-  appraisal (= 0.4.1)
-  aruba (= 0.4.11)
-  bourne (= 1.3.0)
-  bundler (~> 1.1)
-  capybara (= 1.1.2)
-  clearance!
-  cucumber-rails (= 1.1.1)
-  database_cleaner (= 0.8.0)
-  factory_girl_rails (= 3.5.0)
-  psych (~> 1.3.4)
-  rails (= 3.0.20)
-  rspec-rails (= 2.12.2)
-  shoulda-matchers (= 1.2.0)
-  sqlite3 (= 1.3.6)
-  timecop (= 0.3.5)
