--- gemfiles/3.1.11.gemfile
@@ -1,7 +1,0 @@
-# This file was generated by Appraisal
-
-source "http://rubygems.org"
-
-gem "rails", "3.1.11"
-
-gemspec :path=>"../"
