--- gemfiles/3.1.11.gemfile.lock
@@ -1,183 +1,0 @@
-PATH
-  remote: ../
-  specs:
-    clearance (1.0.0.rc5)
-      bcrypt-ruby
-      email_validator
-      rails (>= 3.0)
-
-GEM
-  remote: http://rubygems.org/
-  specs:
-    actionmailer (3.1.11)
-      actionpack (= 3.1.11)
-      mail (~> 2.3.3)
-    actionpack (3.1.11)
-      activemodel (= 3.1.11)
-      activesupport (= 3.1.11)
-      builder (~> 3.0.0)
-      erubis (~> 2.7.0)
-      i18n (~> 0.6)
-      rack (~> 1.3.6)
-      rack-cache (~> 1.2)
-      rack-mount (~> 0.8.2)
-      rack-test (~> 0.6.1)
-      sprockets (~> 2.0.4)
-    activemodel (3.1.11)
-      activesupport (= 3.1.11)
-      builder (~> 3.0.0)
-      i18n (~> 0.6)
-    activerecord (3.1.11)
-      activemodel (= 3.1.11)
-      activesupport (= 3.1.11)
-      arel (~> 2.2.3)
-      tzinfo (~> 0.3.29)
-    activeresource (3.1.11)
-      activemodel (= 3.1.11)
-      activesupport (= 3.1.11)
-    activesupport (3.1.11)
-      multi_json (~> 1.0)
-    appraisal (0.4.1)
-      bundler
-      rake
-    arel (2.2.3)
-    aruba (0.4.11)
-      childprocess (>= 0.2.3)
-      cucumber (>= 1.1.1)
-      ffi (>= 1.0.11)
-      rspec (>= 2.7.0)
-    bcrypt-ruby (3.0.1)
-    bourne (1.3.0)
-      mocha (= 0.13.0)
-    builder (3.0.4)
-    capybara (1.1.2)
-      mime-types (>= 1.16)
-      nokogiri (>= 1.3.3)
-      rack (>= 1.0.0)
-      rack-test (>= 0.5.4)
-      selenium-webdriver (~> 2.0)
-      xpath (~> 0.1.4)
-    childprocess (0.3.8)
-      ffi (~> 1.0, >= 1.0.11)
-    cucumber (1.2.1)
-      builder (>= 2.1.2)
-      diff-lcs (>= 1.1.3)
-      gherkin (~> 2.11.0)
-      json (>= 1.4.6)
-    cucumber-rails (1.1.1)
-      capybara (>= 1.1.1)
-      cucumber (>= 1.1.0)
-      nokogiri (>= 1.5.0)
-    database_cleaner (0.8.0)
-    diff-lcs (1.1.3)
-    email_validator (1.3.0)
-      activemodel
-    erubis (2.7.0)
-    factory_girl (3.5.0)
-      activesupport (>= 3.0.0)
-    factory_girl_rails (3.5.0)
-      factory_girl (~> 3.5.0)
-      railties (>= 3.0.0)
-    ffi (1.3.1)
-    gherkin (2.11.6)
-      json (>= 1.7.6)
-    hike (1.2.1)
-    i18n (0.6.1)
-    json (1.7.7)
-    mail (2.3.3)
-      i18n (>= 0.4.0)
-      mime-types (~> 1.16)
-      treetop (~> 1.4.8)
-    metaclass (0.0.1)
-    mime-types (1.21)
-    mocha (0.13.0)
-      metaclass (~> 0.0.1)
-    multi_json (1.6.0)
-    nokogiri (1.5.6)
-    polyglot (0.3.3)
-    psych (1.3.4)
-    rack (1.3.10)
-    rack-cache (1.2)
-      rack (>= 0.4)
-    rack-mount (0.8.3)
-      rack (>= 1.0.0)
-    rack-ssl (1.3.3)
-      rack
-    rack-test (0.6.2)
-      rack (>= 1.0)
-    rails (3.1.11)
-      actionmailer (= 3.1.11)
-      actionpack (= 3.1.11)
-      activerecord (= 3.1.11)
-      activeresource (= 3.1.11)
-      activesupport (= 3.1.11)
-      bundler (~> 1.0)
-      railties (= 3.1.11)
-    railties (3.1.11)
-      actionpack (= 3.1.11)
-      activesupport (= 3.1.11)
-      rack-ssl (~> 1.3.2)
-      rake (>= 0.8.7)
-      rdoc (~> 3.4)
-      thor (~> 0.14.6)
-    rake (10.0.3)
-    rdoc (3.12.1)
-      json (~> 1.4)
-    rspec (2.12.0)
-      rspec-core (~> 2.12.0)
-      rspec-expectations (~> 2.12.0)
-      rspec-mocks (~> 2.12.0)
-    rspec-core (2.12.2)
-    rspec-expectations (2.12.1)
-      diff-lcs (~> 1.1.3)
-    rspec-mocks (2.12.2)
-    rspec-rails (2.12.2)
-      actionpack (>= 3.0)
-      activesupport (>= 3.0)
-      railties (>= 3.0)
-      rspec-core (~> 2.12.0)
-      rspec-expectations (~> 2.12.0)
-      rspec-mocks (~> 2.12.0)
-    rubyzip (0.9.9)
-    selenium-webdriver (2.29.0)
-      childprocess (>= 0.2.5)
-      multi_json (~> 1.0)
-      rubyzip
-      websocket (~> 1.0.4)
-    shoulda-matchers (1.2.0)
-      activesupport (>= 3.0.0)
-    sprockets (2.0.4)
-      hike (~> 1.2)
-      rack (~> 1.0)
-      tilt (~> 1.1, != 1.3.0)
-    sqlite3 (1.3.6)
-    thor (0.14.6)
-    tilt (1.3.3)
-    timecop (0.3.5)
-    treetop (1.4.12)
-      polyglot
-      polyglot (>= 0.3.1)
-    tzinfo (0.3.35)
-    websocket (1.0.7)
-    xpath (0.1.4)
-      nokogiri (~> 1.3)
-
-PLATFORMS
-  ruby
-
-DEPENDENCIES
-  appraisal (= 0.4.1)
-  aruba (= 0.4.11)
-  bourne (= 1.3.0)
-  bundler (~> 1.1)
-  capybara (= 1.1.2)
-  clearance!
-  cucumber-rails (= 1.1.1)
-  database_cleaner (= 0.8.0)
-  factory_girl_rails (= 3.5.0)
-  psych (~> 1.3.4)
-  rails (= 3.1.11)
-  rspec-rails (= 2.12.2)
-  shoulda-matchers (= 1.2.0)
-  sqlite3 (= 1.3.6)
-  timecop (= 0.3.5)
