--- gemfiles/3.2.12.gemfile
@@ -1,7 +1,0 @@
-# This file was generated by Appraisal
-
-source "http://rubygems.org"
-
-gem "rails", "3.2.12"
-
-gemspec :path=>"../"
