--- gemfiles/3.2.13.rc2.gemfile
@@ -1,7 +1,0 @@
-# This file was generated by Appraisal
-
-source "http://rubygems.org"
-
-gem "rails", "3.2.13.rc2"
-
-gemspec :path=>"../"
