+++ lib/generators/clearance/specs/templates/features/clearance/user_signs_out_spec.rb
