+++ lib/generators/clearance/specs/templates/features/clearance/visitor_signs_in_spec.rb
