+++ lib/generators/clearance/specs/templates/features/clearance/visitor_signs_up_spec.rb
