+++ lib/generators/clearance/specs/templates/features/clearance/visitor_updates_password_spec.rb
